+++ snx-rs-gui/src/settings.rs
@@ -76,6 +76,7 @@ struct MyWidgets {
     ike_port: gtk::Entry,
     ike_persist: gtk::CheckButton,
     ike_transport: gtk::ComboBoxText,
+    esp_transport: gtk::ComboBoxText,
     no_keepalive: gtk::CheckButton,
     icon_theme: gtk::ComboBoxText,
     error: gtk::Label,
@@ -277,6 +278,7 @@ impl SettingsDialog {
         let esp_lifetime = gtk::Entry::builder()
             .text(params.esp_lifetime.as_secs().to_string())
             .build();
+        let esp_transport = gtk::ComboBoxText::builder().build();
         let ike_port = gtk::Entry::builder().text(params.ike_port.to_string()).build();
         let ike_persist = gtk::CheckButton::builder().active(params.ike_persist).build();
         let ike_transport = gtk::ComboBoxText::builder().build();
@@ -418,6 +420,7 @@ impl SettingsDialog {
             ca_cert,
             ike_lifetime,
             esp_lifetime,
+            esp_transport,
             ike_port,
             ike_persist,
             ike_transport,
@@ -554,6 +557,7 @@ impl SettingsDialog {
             .collect();
         params.ike_lifetime = Duration::from_secs(self.widgets.ike_lifetime.text().parse()?);
         params.esp_lifetime = Duration::from_secs(self.widgets.esp_lifetime.text().parse()?);
+        params.esp_transport = self.widgets.esp_transport.active().unwrap_or_default().into();
         params.ike_port = self.widgets.ike_port.text().parse()?;
         params.ike_persist = self.widgets.ike_persist.is_active();
         params.no_keepalive = self.widgets.no_keepalive.is_active();
@@ -649,6 +653,17 @@ impl SettingsDialog {
         ike_transport_box
     }
 
+    fn esp_transport_box(&self) -> gtk::Box {
+        let esp_transport_box = self.form_box("ESP transport");
+        self.widgets.esp_transport.insert_text(0, "UDP");
+        self.widgets.esp_transport.insert_text(1, "TCPT");
+        self.widgets
+            .esp_transport
+            .set_active(Some(self.params.esp_transport.as_u32()));
+        esp_transport_box.pack_start(&self.widgets.esp_transport, false, true, 0);
+        esp_transport_box
+    }
+
     fn user_box(&self) -> gtk::Box {
         let user_box = self.form_box("User name");
         user_box.pack_start(&self.widgets.user_name, false, true, 0);
@@ -742,6 +757,9 @@ impl SettingsDialog {
         let esp_lifetime = self.form_box("ESP lifetime, seconds");
         esp_lifetime.pack_start(&self.widgets.esp_lifetime, false, true, 0);
         misc_box.pack_start(&esp_lifetime, false, true, 6);
+
+        let esp_transport_box = self.esp_transport_box();
+        misc_box.pack_start(&esp_transport_box, false, true, 6);
 
         let ike_port = self.form_box("IKE port");
         ike_port.pack_start(&self.widgets.ike_port, false, true, 0);