+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## v2.10.0 (unreleased)
-- Added support for TCP Transport mode (TCPT) in IKE via the `ike-transport=tcpt` option. Use this option if you encounter the "failed to fill whole buffer" error.
+## v3.0.0 (unreleased)
+- Added support for TCP Transport mode (TCPT) for IPSec tunnels. It can be selected by two options: `ike-transport=tcpt` and `esp-transport=tcpt`.
+  Use the first option only if you encounter the "failed to fill whole buffer" error. Use the second option only if you encounter the "Probing failed, server is not reachable via ESPinUDP tunnel" error.
 - Implemented a sign-out call after tunnel disconnection.
 - Resolved a compatibility issue with OpenSSL 1.x when building on older hosts.